# backend/hotelflow_server/swagger.yaml
openapi: 3.0.0
info:
  title: HotelFlow API Documentation
  description: |
    Complete API documentation for the HotelFlow booking system.
    This API provides endpoints for hotels, rooms, bookings, services, invoices, guests, images, RFID keys, tags, and super admin management.
  version: 1.0.0
  contact:
    name: HotelFlow API Support
servers:
  - url: /api
    description: Current server (auto-detected)
  - url: http://172.16.50.35:8000/api
    description: Development server
  - url: http://localhost:8000/api
    description: Local server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Laravel Sanctum token authentication
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [user, hotel, super_admin]
        isVerified:
          type: boolean
        two_factor_enabled:
          type: boolean
        tax_number:
          type: string
          nullable: true
        bank_account:
          type: string
          nullable: true
        eu_tax_number:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        message:
          type: string
        errors:
          type: object
paths:
...
  /auth/register-user:
    post:
      tags:
        - Authentication
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/register-hotel:
    post:
      tags:
        - Authentication
      summary: Register a new hotel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
                - location
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                location:
                  type: string
      responses:
        '201':
          description: Hotel registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: Validation error
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                two_factor_code:
                  type: string
                  description: Required if 2FA is enabled
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  email:
                    type: string
                  role:
                    type: string
                  token:
                    type: string
                  requires_2fa:
                    type: boolean
        '401':
          description: Invalid credentials
        '403':
          description: Email not verified
  /auth/logout:
    post:
      tags:
        - Authentication
      security:
        - bearerAuth: []
      summary: Logout user
      responses:
        '200':
          description: Logout successful
  /auth/verify-email/{token}:
    get:
      tags:
        - Authentication
      summary: Verify email address
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Email verified successfully
        '400':
          description: Invalid or expired token
  /auth/resend-verification:
    post:
      tags:
        - Authentication
      summary: Resend verification email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Verification email sent
  /auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Password reset email sent
  /auth/reset-password:
    post:
      tags:
        - Authentication
      summary: Reset password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - email
                - password
              properties:
                token:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Password reset successful
        '400':
          description: Invalid token
  /auth/user/{id}:
    get:
      tags:
        - Authentication
      security:
        - bearerAuth: []
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
  /auth/me:
    get:
      tags:
        - Authentication
      security:
        - bearerAuth: []
      summary: Get current authenticated user
      responses:
        '200':
          description: Current user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /auth/updateuser/{id}:
    put:
      tags:
        - Authentication
      security:
        - bearerAuth: []
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /auth/deleteuser/{id}:
    post:
      tags:
        - Authentication
      security:
        - bearerAuth: []
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User deleted successfully
        '404':
          description: User not found
  /auth/admin/profile:
    put:
      tags:
        - Authentication
      security:
        - bearerAuth: []
      summary: Update admin profile (including invoice fields)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                tax_number:
                  type: string
                bank_account:
                  type: string
                eu_tax_number:
                  type: string
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /auth/2fa/enable:
    post:
      tags:
        - Authentication
      security:
        - bearerAuth: []
      summary: Enable 2FA
      responses:
        '200':
          description: 2FA setup initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  qr_code_url:
                    type: string
                  secret:
                    type: string
  /auth/2fa/verify-enable:
    post:
      tags:
        - Authentication
      security:
        - bearerAuth: []
      summary: Verify and enable 2FA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
              properties:
                code:
                  type: string
      responses:
        '200':
          description: 2FA enabled successfully
        '400':
          description: Invalid code
  /auth/2fa/disable:
    post:
      tags:
        - Authentication
      security:
        - bearerAuth: []
      summary: Disable 2FA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
              properties:
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: 2FA disabled successfully
  /auth/verify-2fa:
    post:
      tags:
        - Authentication
      security:
        - bearerAuth: []
      summary: Verify 2FA code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
              properties:
                code:
                  type: string
      responses:
        '200':
          description: Code verified successfully
        '400':
          description: Invalid code
  /auth/2fa/recovery/request:
    post:
      tags:
        - Authentication
      summary: Request 2FA recovery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Recovery email sent
  /auth/2fa/recovery/confirm:
    post:
      tags:
        - Authentication
      summary: Confirm 2FA recovery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - email
              properties:
                token:
                  type: string
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Recovery confirmed, 2FA disabled

  # ========== HOTEL ENDPOINTS ==========
  /hotels:
    get:
      tags:
        - Hotels
      summary: Get all hotels
      description: Returns all hotels. If authenticated as hotel admin, returns only their hotels.
      responses:
        '200':
          description: List of hotels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Hotel'
    post:
      tags:
        - Hotels
      security:
        - bearerAuth: []
      summary: Create a new hotel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - location
              properties:
                name:
                  type: string
                location:
                  type: string
                description:
                  type: string
                type:
                  type: string
                  enum: [hotel, apartment, villa, other]
                starRating:
                  type: integer
                  minimum: 1
                  maximum: 5
      responses:
        '201':
          description: Hotel created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hotel'
  /hotels/{id}:
    get:
      tags:
        - Hotels
      summary: Get hotel by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Hotel details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hotel'
        '404':
          description: Hotel not found
  /hotels/upgrade/{id}:
    put:
      tags:
        - Hotels
      security:
        - bearerAuth: []
      summary: Upgrade hotel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                starRating:
                  type: integer
                  minimum: 1
                  maximum: 5
      responses:
        '200':
          description: Hotel upgraded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hotel'
  /hotels/{id}/cover-image:
    post:
      tags:
        - Hotels
      security:
        - bearerAuth: []
      summary: Upload hotel cover image
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - image
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Image uploaded successfully
        '404':
          description: Hotel not found
  /hotels/{id}/billing-info:
    get:
      tags:
        - Hotels
      security:
        - bearerAuth: []
      summary: Get hotel billing information
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Billing information
          content:
            application/json:
              schema:
                type: object
    put:
      tags:
        - Hotels
      security:
        - bearerAuth: []
      summary: Update hotel billing information
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tax_number:
                  type: string
                bank_account:
                  type: string
                eu_tax_number:
                  type: string
      responses:
        '200':
          description: Billing information updated
  /hotels/delete/{id}:
    delete:
      tags:
        - Hotels
      security:
        - bearerAuth: []
      summary: Delete hotel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Hotel deleted successfully
        '404':
          description: Hotel not found
...
  /hotels/{hotel}/tags:
    post:
      tags:
        - Hotels
      security:
        - bearerAuth: []
      summary: Add tag to hotel
      parameters:
        - name: hotel
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tag_id
              properties:
                tag_id:
                  type: integer
      responses:
        '200':
          description: Tag added successfully
  /hotels/{hotel}/tags/{tag}:
    delete:
      tags:
        - Hotels
      security:
        - bearerAuth: []
      summary: Remove tag from hotel
      parameters:
        - name: hotel
          in: path
          required: true
          schema:
            type: integer
        - name: tag
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Tag removed successfully

  # ========== ROOM ENDPOINTS ==========
  /rooms/hotel/{hotel_id}:
    get:
      tags:
        - Rooms
      summary: Get rooms by hotel ID
      parameters:
        - name: hotel_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
  /rooms/{id}:
    get:
      tags:
        - Rooms
      summary: Get room by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Room details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '404':
          description: Room not found
...
  /rooms/update/{id}:
    put:
      tags:
        - Rooms
      security:
        - bearerAuth: []
      summary: Update room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                capacity:
                  type: integer
                price:
                  type: number
                  format: float
      responses:
        '200':
          description: Room updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
...